requirements.txt
groupyapi
boto3